feat: add support for hiding resources (Resolves #547)

+++ cmd/delete-all.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"context"
 	"fmt"
+	"github.com/elasticpath/epcc-cli/config"
 	"github.com/elasticpath/epcc-cli/external/aliases"
 	"github.com/elasticpath/epcc-cli/external/apihelper"
 	"github.com/elasticpath/epcc-cli/external/httpclient"
@@ -34,7 +35,15 @@ func NewDeleteAllCommand(parentCmd *cobra.Command) func() {
 		},
 	}
 
+	e := config.GetEnv()
+	hiddenResources := map[string]struct{}{}
+
+	for _, v := range e.EPCC_CLI_DISABLE_RESOURCES {
+		hiddenResources[v] = struct{}{}
+	}
+
 	for _, resource := range resources.GetPluralResources() {
+
 		if resource.GetCollectionInfo == nil {
 			continue
 		}
@@ -42,6 +51,17 @@ func NewDeleteAllCommand(parentCmd *cobra.Command) func() {
 		if resource.DeleteEntityInfo == nil {
 			continue
 		}
+
+		if _, ok := hiddenResources[resource.SingularName]; ok {
+			log.Tracef("Hiding resource %s", resource.SingularName)
+			continue
+		}
+
+		if _, ok := hiddenResources[resource.PluralName]; ok {
+			log.Tracef("Hiding resource %s", resource.SingularName)
+			continue
+		}
+
 		resourceName := resource.PluralName
 
 		var deleteAllResourceCmd = &cobra.Command{
